+++ src/mas/ast.rs
@@ -332,13 +332,13 @@ fn assert_reg(reg: Reg) {
 }
 
 fn assert_offs(offs: i32, width: u8) {
-    let mask = (-1 as u32) << (width as uint);
+    let mask = (-1 as u32) << (width as usize);
     // Check that the sign bit is extended all the way.
     assert_eq!((mask & (offs as u32) == mask),
-               (((offs as u32) & (1<<((width-1) as uint)) != 0)));
+               (((offs as u32) & (1<<((width-1) as usize)) != 0)));
     // No sign bit means none of the high bits should be set.
     assert_eq!((mask & (offs as u32) == 0),
-               ((offs as u32) & (1<<((width-1) as uint))) == 0);
+               ((offs as u32) & (1<<((width-1) as usize))) == 0);
 }
 
 // Helper functions that will make instruction representations, and also
